Integrate dhcp_relay config plugin
